refactor: 결제/대시보드 BigDecimal 기반 리팩토링 및 통계 개선

+++ src/main/java/com/byunsum/ticket_reservation/payment/dto/PaymentResponse.java
@@ -5,6 +5,7 @@ import com.byunsum.ticket_reservation.payment.domain.PaymentMethod;
 import com.byunsum.ticket_reservation.payment.domain.PaymentStatus;
 import io.swagger.v3.oas.annotations.media.Schema;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public class PaymentResponse {
@@ -12,16 +13,16 @@ public class PaymentResponse {
     private Long id;
 
     @Schema(description = "총 결제 금액")
-    private int amount;
+    private BigDecimal amount;
 
     @Schema(description = "부분 취소된 금액 (없으면 null)")
-    private Integer partialAmount;
+    private BigDecimal partialAmount;
 
     @Schema(description = "취소 수수료 (없으면 null)")
-    private Integer cancelFee;
+    private BigDecimal cancelFee;
 
     @Schema(description = "실제 환불된 금액 (없으면 null)")
-    private Integer refundAmount;
+    private BigDecimal refundAmount;
 
     @Schema(description = "결제 수단")
     private PaymentMethod paymentMethod;
@@ -41,7 +42,7 @@ public class PaymentResponse {
     @Schema(description = "계좌번호")
     private String accountNumber;
 
-    public PaymentResponse(Long id, int amount, Integer partialAmount, Integer cancelFee, Integer refundAmount, PaymentMethod paymentMethod, PaymentStatus paymentStatus, LocalDateTime createdAt, LocalDateTime cancelledAt, PaymentCancelReason cancelReason, String accountNumber) {
+    public PaymentResponse(Long id, BigDecimal amount, BigDecimal partialAmount, BigDecimal cancelFee, BigDecimal refundAmount, PaymentMethod paymentMethod, PaymentStatus paymentStatus, LocalDateTime createdAt, LocalDateTime cancelledAt, PaymentCancelReason cancelReason, String accountNumber) {
         this.id = id;
         this.amount = amount;
         this.partialAmount = partialAmount;
@@ -59,19 +60,19 @@ public class PaymentResponse {
         return id;
     }
 
-    public int getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
-    public Integer getPartialAmount() {
+    public BigDecimal getPartialAmount() {
         return partialAmount;
     }
 
-    public Integer getCancelFee() {
+    public BigDecimal getCancelFee() {
         return cancelFee;
     }
 
-    public Integer getRefundAmount() {
+    public BigDecimal getRefundAmount() {
         return refundAmount;
     }
